@@ -20,7 +20,7 @@ GDS::Application.configure do
   # config.action_dispatch.rack_cache = true
 
   # Disable Rails's static asset server (Apache or nginx will already do this).
-  config.serve_static_assets =  true
+  config.serve_static_files =  true
 
   # Compress JavaScripts and CSS.
   config.assets.js_compressor = :uglifier
