@@ -1,12 +1,3 @@
-# This file should contain all the record creation needed to seed the database with its default values.
-# The data can then be loaded with the rake db:seed (or created alongside the db with db:setup).
-#
-# Examples:
-#
-#   cities = City.create([{ name: 'Chicago' }, { name: 'Copenhagen' }])
-#   Mayor.create(name: 'Emanuel', city: cities.first)
-
-
 Role.delete_all
 Role.create!(name: 'user', level: 1)
 Role.create!(name: 'proband', level: 1)
@@ -20,8 +11,8 @@ ArgumentType.delete_all
 ArgumentType.create!(name: 'proband')
 ArgumentType.create!(name: 'moderator')
 ArgumentType.create!(name: 'observer')
-/
 
+=begin example data
 ResearchInstitute.delete_all
 ResearchInstitute.create!(
 	[
@@ -81,4 +72,4 @@ Company.create!(
 			research_institute: ResearchInstitute.take
 		}
 	])
-/
+=end
