@@ -7,14 +7,17 @@
 #    t.string   "remember_token"
 #    t.integer  "role_id"
 
+# TODO different roles
+
 FactoryGirl.define do
   factory :user do
 	  sequence(:firstname)  { |n| "User #{n}" }
     sequence(:username)  { |n| "User #{n}" }
     sequence(:lastname)  { |n| "the #{n} of his kind" }
-    sequence(:email) { |n| "user_#{n}@example.com" } 
+    sequence(:email) { |n| "user_#{n}@example.com" }
     password "123456789"
     password_confirmation "123456789"
-    role { Role.where(name:"user").first }
+    confirmed_at Date.yesterday
+    role { Role.where(name:"moderator").first }
   end
 end
