@@ -1,81 +1,22 @@
-require 'rubygems'
-require 'spork'
-#uncomment the following line to use spork with the debugger
-#require 'spork/ext/ruby-debug'
+ENV["RAILS_ENV"] ||= 'test'
 
-Spork.prefork do
-  # Loading more in this block will cause your tests to run faster. However,
-  # if you change any configuration or code from libraries loaded here, you'll
-  # need to restart spork for it take effect.
+require File.expand_path("../../config/environment", __FILE__)
+require File.expand_path("../../db/seeds", __FILE__)
+require 'rspec/rails'
+require 'factory_girl_rails'
 
-  ENV["RAILS_ENV"] ||= 'test'
-  require File.expand_path("../../config/environment", __FILE__)
-  require File.expand_path("../../db/seeds", __FILE__)
-  require 'rspec/rails'
-  require 'rspec/autorun'
-  require 'factory_girl_rails'
-  require 'simplecov'
+Dir[Rails.root.join("spec/support/**/*.rb")].each {|f| require f}
+Rails.application.routes.default_url_options[:host] = 'localhost:5000'
+# ActiveRecord::Migration.check_pending! if defined?(ActiveRecord::Migration)
 
-  #SimpleCov.start do
-  #  add_filter "/spec/"
-  #end
-
-  # Requires supporting ruby files with custom matchers and macros, etc,
-  # in spec/support/ and its subdirectories.
-  Dir[Rails.root.join("spec/support/**/*.rb")].each {|f| require f}
-
-  # Checks for pending migrations before tests are run.
-  # If you are not using ActiveRecord, you can remove this line.
-  ActiveRecord::Migration.check_pending! if defined?(ActiveRecord::Migration)
-
-
-  RSpec.configure do |config|
-    # ## Mock Framework
-    #
-    # If you prefer to use mocha, flexmock or RR, uncomment the appropriate line:
-    #
-    # config.mock_with :mocha
-    # config.mock_with :flexmock
-    # config.mock_with :rr
-
-    # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
-    config.fixture_path = "#{::Rails.root}/spec/fixtures"
-
-    # If you're not using ActiveRecord, or you'd prefer not to run each of your
-    # examples within a transaction, remove the following line or assign false
-    # instead of true.
-    config.use_transactional_fixtures = true
-
-    # If true, the base class of anonymous controllers will be inferred
-    # automatically. This will be the default behavior in future versions of
-    # rspec-rails.
-    config.infer_base_class_for_anonymous_controllers = false
-
-    # Run specs in random order to surface order dependencies. If you find an
-    # order dependency and want to debug it, you can fix the order by providing
-    # the seed, which is printed after each run.
-    #     --seed 1234
-    config.order = "random"
-    config.include Capybara::DSL
-    require 'capybara/rails'
-    class ActionDispatch::IntegrationTest
-      include Capybara::DSL
-      def teardown
-        Capybara.reset_sessions!
-        Capybara.use_default_driver
-      end
-    end
-
-    include Devise::TestHelpers
-
-  end
-
-
+RSpec.configure do |config|
+  # config.mock_with :rr
+  config.fixture_path = "#{::Rails.root}/spec/fixtures"
+  config.use_transactional_fixtures = false
+  config.infer_base_class_for_anonymous_controllers = false
+  config.order = "random"
+  config.include FactoryGirl::Syntax::Methods
+  config.include Devise::TestHelpers, type: :controller
+  config.include Rails.application.routes.url_helpers
 end
-
-Spork.each_run do
-  # This code will be run each time you run your specs.
-
-end
-
 
